refactor: add namespace CommandLineOption
