+++ CMakeLists.txt
@@ -30,6 +30,7 @@ include_directories(
 set(SOURCE_FILES
 	"src/app/AppSetting.cpp"
 	"src/app/CommandLine.cpp"
+	"src/app/Option.cpp"
 
 	"src/dynamics/WindModel.cpp"
